+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/module/installer/bundle/listener/GuiceyStartupListenerAdapter.java
@@ -1,6 +1,7 @@
 package ru.vyarus.dropwizard.guice.module.installer.bundle.listener;
 
 import com.google.common.base.Throwables;
+import io.dropwizard.core.Configuration;
 import ru.vyarus.dropwizard.guice.module.lifecycle.GuiceyLifecycleAdapter;
 import ru.vyarus.dropwizard.guice.module.lifecycle.event.run.ApplicationRunEvent;
 
@@ -9,16 +10,16 @@ import ru.vyarus.dropwizard.guice.module.lifecycle.event.run.ApplicationRunEvent
  *
  * @author Vyacheslav Rusakov
  * @since 28.09.2019
+ * @param <C> configuration type
  */
-public class GuiceyStartupListenerAdapter extends GuiceyLifecycleAdapter {
-    private final GuiceyStartupListener listener;
+public class GuiceyStartupListenerAdapter<C extends Configuration> extends GuiceyLifecycleAdapter {
+    private final GuiceyStartupListener<C> listener;
 
-    public GuiceyStartupListenerAdapter(final GuiceyStartupListener listener) {
+    public GuiceyStartupListenerAdapter(final GuiceyStartupListener<C> listener) {
         this.listener = listener;
     }
 
     @Override
-    @SuppressWarnings("unchecked")
     protected void applicationRun(final ApplicationRunEvent event) {
         try {
             listener.configure(event.getConfiguration(), event.getEnvironment(), event.getInjector());